Enhance step selection
