+++ src/index.html
@@ -7,19 +7,6 @@
     <meta name="viewport" content="width=device-width,initial-scale=1">
 </head>
 <body>
-    <style type="text/css">
-        #root {
-            border: 1px solid grey;
-            margin: auto;
-            font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto, "PingFang SC", "Hiragino Sans GB", "Microsoft YaHei", "Helvetica Neue", Helvetica, Arial, sans-serif;
-            font-size: 12px;
-            line-height: 1.5;
-            color: #666;
-        }
-    </style>
-
-    <div id="root">
-        <div id="bookbuilder"></div>
-    </div>
+    <div id="bookbuilder"></div>
 </body>
 </html>
