+++ src/js/components/App.js
@@ -1,35 +1,31 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import Steps from 'antd/lib/steps';
+import Steps, { Step } from 'antd/lib/steps';
+import { bindActionCreators } from 'redux';
+import { chooseStep } from '../actions/index';
 import BookBinding from './BookBinding';
 import Color from './Color';
 import Format from './Format';
-import Footer from './Footer';
-import Summary from './Summary';
+import Payment from './Payment';
 import Finish from './Finish';
 import Error from './Error';
-
-const Step = Steps.Step;
 
 class App extends Component {
     steps() {
         return (<Steps size="small" current={this.props.currentStep}>
-            {this.props.steps.map(item => <Step key={item.title} title={item.title} />)}
+            {this.props.steps.map(item => <Step key={item.title} title={item.title} onClick={() => item.step < this.props.currentStep && this.props.chooseStep(item.step)} />)}
         </Steps>);
     }
 
     render() {
-        return (<div>
+        return (<div className="app">
             <div className="steps">{ this.steps() }</div>
-            <div className="content">
-                { this.props.currentStep === 0 ? <BookBinding /> : null }
-                { this.props.currentStep === 1 ? <Color /> : null }
-                { this.props.currentStep === 2 ? <Format /> : null }
-                { this.props.currentStep === 3 ? <Summary /> : null }
-                { this.props.currentStep === 4 ? <Finish /> : null }
-                { this.props.currentStep === 5 ? <Error /> : null }
-            </div>
-            { this.props.currentStep !== this.props.steps.length ? <Footer /> : null }
+            { this.props.currentStep === 0 ? <BookBinding /> : null }
+            { this.props.currentStep === 1 ? <Color /> : null }
+            { this.props.currentStep === 2 ? <Format /> : null }
+            { this.props.currentStep === 3 ? <Payment /> : null }
+            { this.props.currentStep === 4 ? <Finish /> : null }
+            { this.props.currentStep === 5 ? <Error /> : null }
         </div>);
     }
 }
@@ -41,4 +37,8 @@ function mapStateToProps(state) {
     };
 }
 
-export default connect(mapStateToProps)(App);
+function mapDispatchToProps(dispatch) {
+    return bindActionCreators({ chooseStep }, dispatch);
+}
+
+export default connect(mapStateToProps, mapDispatchToProps)(App);
